refactor(phase-48.1): split multi_molecule_writer.py from gromacs_writer (no behavior change)
- Move write_multi_molecule_gro_file + write_multi_molecule_top_file to multi_molecule_writer.py (~464 lines)
- _registry singleton imported from _shared (moved in Wave 1)
- NO try/except added to GRO writer (intentionally absent — research §3)
- custom_by_moltype dict pattern preserved (NOT unified with interface's if custom_guest_info:)
- 14 byte-equivalence + hydrate grompp tests pass
